move files to sync subdirectory
