+++ sync/waitgroup_test.go
@@ -4,7 +4,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package debugsync
+package sync
 
 import (
 	"sync/atomic"
